21.12.2020 christmas

+++ cub3d.h
@@ -6,7 +6,7 @@
 /*   By: dkarthus <dkarthus@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/12/04 15:03:47 by dkarthus          #+#    #+#             */
-/*   Updated: 2020/12/20 19:08:44 by dkarthus         ###   ########.fr       */
+/*   Updated: 2020/12/21 18:23:25 by dkarthus         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -57,11 +57,11 @@ typedef struct	s_txtr
 
 typedef struct	s_obj
 {
-	float x;
-	float y;
-	float dir;
-	float fov_st;
-	float fov_end;
+	double x;
+	double y;
+	double dir;
+	double fov_st;
+	double fov_end;
 }				t_obj;
 
 typedef struct	s_vars
